requirements/local.txt
-r base.txt

Werkzeug==1.0.1 # https://github.com/pallets/werkzeug
ipdb==0.13.3  # https://github.com/gotcha/ipdb
psycopg2==2.8.5 --no-binary psycopg2  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
mypy==0.770  # https://github.com/python/mypy
django-stubs==1.5.0  # https://github.com/typeddjango/django-stubs
pytest==6.0.1  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.4  # https://github.com/Frozenball/pytest-sugar

# Tools
django-extensions==2.2.9

# Documentation
# ------------------------------------------------------------------------------
sphinx==3.2.0  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==0.7.1 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8==3.8.3  # https://github.com/PyCQA/flake8
pylint-django==2.3.0  # https://github.com/PyCQA/pylint-django
pre-commit==2.6.0  # https://github.com/pre-commit/pre-commit

# librerias
drf-nested-routers
djangorestframework-filters==1.0.0.dev0
drf-flex-fields
pyjwt==1.7.1